+++ mobile/src/tabs/discover/index.js
@@ -1,93 +1,57 @@
 import React from 'react';
 import { Query } from 'react-apollo';
-import { Button, FlatList } from 'react-native';
 import { createStackNavigator } from 'react-navigation';
-import PropTypes from 'prop-types';
-import { get } from 'lodash';
 
+import FeedView from 'ui/FeedView';
 import BackgroundView from 'ui/BackgroundView';
-import { ErrorCard } from 'ui/Card';
+import TileContentFeed from './tileContentFeed';
 import tabBarIcon from '../tabBarIcon';
 import GET_DISCOVER_ITEMS from './query';
 
-export class DiscoverScreen extends React.Component {
-  static navigationOptions = {
-    title: 'Discover',
-  };
+const childContentItemLoadingObject = {
+  node: {
+    isLoading: true,
+  },
+};
 
-  static propTypes = {
-    fetchMore: PropTypes.func,
-    keyExtractor: PropTypes.func,
-    ListEmptyComponent: PropTypes.func,
-    navigation: PropTypes.shape({
-      navigate: PropTypes.func,
-    }),
-    onEndReachedThreshold: PropTypes.number,
-  };
-
-  static defaultProps = {
-    fetchMore: undefined,
-    keyExtractor: (item) => item && item.id,
-    onEndReachedThreshold: 0.7,
-  };
-
-  refetchHandler = ({ isLoading, refetch }) =>
-    refetch && ((...args) => !isLoading && refetch(...args));
-
-  fetchMoreHandler = ({ fetchMore, error, isLoading }) =>
-    fetchMore && ((...args) => !isLoading && !error && fetchMore(...args));
-
-  render() {
-    const {
-      fetchMore,
-      keyExtractor,
-      ListEmptyComponent,
-      onEndReachedThreshold,
-    } = this.props;
-    return (
-      <BackgroundView>
-        <Query query={GET_DISCOVER_ITEMS}>
-          {({ loading, error, data, refetch }) => {
-            if (loading) return 'Loading...';
-
-            return (
-              <FlatList
-                data={get(data, 'contentChannels', [])}
-                keyExtractor={keyExtractor}
-                ListEmptyComponent={
-                  error && !loading && (!data || !data.length) ? (
-                    <ErrorCard error={error} />
-                  ) : (
-                    ListEmptyComponent
-                  )
-                }
-                renderItem={({ item }) => (
-                  <Button
-                    title={item.name}
-                    onPress={() => {
-                      this.props.navigation.navigate('ContentFeed', {
-                        itemId: item.id,
-                        itemTitle: item.name,
-                      });
-                    }}
-                  />
-                )}
-                onEndReached={this.fetchMoreHandler({
-                  fetchMore,
-                  error,
-                  loading,
-                })}
-                onEndReachedThreshold={onEndReachedThreshold}
-                onRefresh={this.refetchHandler({ loading, refetch })}
-                refreshing={loading}
-              />
-            );
+const DiscoverScreen = () => (
+  <BackgroundView>
+    <Query query={GET_DISCOVER_ITEMS}>
+      {({ error, loading, data: { contentChannels = [] } = {}, refetch }) => (
+        <FeedView
+          error={error}
+          content={contentChannels}
+          keyExtractor={(item) => item.id}
+          isLoading={loading}
+          refetch={refetch}
+          renderItem={({ item }) => (
+            <TileContentFeed
+              id={item.id}
+              name={item.name}
+              content={item.childContentItemsConnection.edges.map(
+                (edge) => edge.node
+              )}
+            />
+          )}
+          loadingStateObject={{
+            name: '',
+            childContentItemsConnection: {
+              edges: [
+                childContentItemLoadingObject,
+                childContentItemLoadingObject,
+                childContentItemLoadingObject,
+              ],
+            },
           }}
-        </Query>
-      </BackgroundView>
-    );
-  }
-}
+        />
+      )}
+    </Query>
+  </BackgroundView>
+);
+
+DiscoverScreen.navigationOptions = {
+  title: 'Discover',
+};
 
 export const DiscoverStack = createStackNavigator(
   {