+++ mobile/src/tabs/discover/query.js
@@ -5,6 +5,19 @@ const GET_DISCOVER_ITEMS = gql`
     contentChannels {
       id
       name
+      childContentItemsConnection(first: 3) {
+        edges {
+          node {
+            id
+            title
+            coverImage {
+              sources {
+                uri
+              }
+            }
+          }
+        }
+      }
     }
   }
 `;