+++ mobile/src/ui/SideBySideView/Right.js
@@ -1,0 +1,18 @@
+import { View } from 'react-native';
+import { compose } from 'recompose';
+import { enhancer as mediaQuery } from 'ui/MediaQuery';
+import styled from 'ui/styled';
+
+export default compose(
+  mediaQuery(
+    ({ md }) => ({ minWidth: md }),
+    styled(
+      {
+        width: '41.6666666%',
+        height: '100%',
+        overflow: 'hidden',
+      },
+      'SideBySideView.Right'
+    )
+  )
+)(View);