+++ mobile/src/ui/TileImage/index.js
@@ -12,6 +12,9 @@ const CardView = styled(
   ({ theme }) => ({
     borderRadius: theme.sizing.borderRadius,
     overflow: 'hidden',
+    width: '100%',
+    height: '100%',
+    aspectRatio: 1,
   }),
   'TileImage'
 )(View);
@@ -21,6 +24,7 @@ const Title = styled(
     position: 'absolute',
     bottom: theme.sizing.baseUnit,
     left: theme.sizing.baseUnit,
+    right: theme.sizing.baseUnit,
     backgroundColor: theme.colors.transparent,
     color: theme.colors.lightPrimary,
   }),
@@ -32,23 +36,29 @@ const enhance = compose(
   pure
 );
 
-const TileImage = enhance(({ image, link, onPressItem, text, theme }) => (
-  <TouchableWithoutFeedback onPress={() => onPressItem({ ...link })}>
-    <CardView>
-      <GradientOverlayImage
-        source={image}
-        overlayColor={text ? theme.colors.black : null}
-      />
-      {text ? <Title>{text}</Title> : null}
-    </CardView>
-  </TouchableWithoutFeedback>
-));
+const TileImage = enhance(
+  ({ image, link, onPressItem, text, theme, isLoading }) => (
+    <TouchableWithoutFeedback
+      onPress={() => !isLoading && onPressItem({ ...link })}
+    >
+      <CardView>
+        <GradientOverlayImage
+          source={image}
+          isLoading={isLoading}
+          overlayColor={text ? theme.colors.black : null}
+        />
+        <Title isLoading={isLoading}>{text || ''}</Title>
+      </CardView>
+    </TouchableWithoutFeedback>
+  )
+);
 
 TileImage.propTypes = {
   image: GradientOverlayImage.propTypes.source,
-  link: PropTypes.string.isRequired,
+  link: PropTypes.string,
   onPressItem: PropTypes.func,
   text: PropTypes.string,
+  isLoading: PropTypes.bool,
 };
 
 export default TileImage;